+++ ext_emconf.php
@@ -12,7 +12,7 @@ $EM_CONF[$_EXTKEY] = [
             'php' => '7.4.0-0.0.0',
             'typo3' => '10.4.11-11.5.99',
             'form' => '10.4.11-11.5.99',
-            'jobrouter_base' => '1.2.0-1.99.99',
+            'jobrouter_base' => '1.4.0-1.99.99',
             'jobrouter_connector' => '1.0.0-1.99.99',
         ],
         'conflicts' => [],